backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
google-api-python-client==2.126.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.7.1
aiofiles==23.2.1
httpx==0.27.0
pytz==2024.1
requests==2.31.0
instagrapi==2.1.2
linkedin-api==2.1.1
